+++ src/main/java/org/komamitsu/fluency/buffer/Buffer.java
@@ -61,7 +61,7 @@ public abstract class Buffer<T extends Buffer.Config>
 
     public abstract long getAllocatedSize();
 
-    public int getMaxSize()
+    public long getMaxSize()
     {
         return bufferConfig.getMaxBufferSize();
     }
@@ -73,15 +73,15 @@ public abstract class Buffer<T extends Buffer.Config>
 
     public abstract static class Config<T extends Buffer, C extends Config>
     {
-        protected int maxBufferSize = 512 * 1024 * 1024;
+        protected long maxBufferSize = 512 * 1024 * 1024;
         protected boolean ackResponseMode = false;
 
-        public int getMaxBufferSize()
+        public long getMaxBufferSize()
         {
             return maxBufferSize;
         }
 
-        public C setMaxBufferSize(int maxBufferSize)
+        public C setMaxBufferSize(long maxBufferSize)
         {
             this.maxBufferSize = maxBufferSize;
             return (C)this;
